% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Function cleans data}
\usage{
clean(data)
}
\arguments{
\item{data}{The data set being cleaned}
}
\value{
Data frame without NA values
}
\description{
Use this function to quickly remove NA values from data.
Returns data without NA.
It is suggested to assign this to a new variable.
}
