% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.color.location.R, R/quick_crab_plot.R
\name{crab_plot}
\alias{crab_plot}
\title{Function Plots Crab Data}
\usage{
crab_plot(data, columnx, columny)

crab_plot(data, columnx, columny)
}
\arguments{
\item{data}{The data set crabs}

\item{columnx}{selected column for x axis}

\item{columny}{selected column for y axis}

\item{columnz}{selected column for color of data}

\item{column}{Column for X value}

\item{column1}{Column for Y value}
}
\value{
plot The plot visualizing three crab characteristics

X-Y plot
}
\description{
Use this function to plot three columns of data using the crabs data set.
This plot will show the x and y axis, along with a column selected for data coloring.
With this plot function, three crab characteristics can be visualized.

Use this plot to quickly plot two numeric columns to view data correlation.
This plot is useful for fast visualization for comparisons.
}
