% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crabs_xysl.R
\name{crab_xysl}
\alias{crab_xysl}
\title{Function Plots Crab Data}
\usage{
crab_xysl(crab_data, columnx, columny)
}
\arguments{
\item{columnx}{selected column for x axis}

\item{columny}{selected column for y axis}

\item{data}{data frame being used}
}
\value{
plot The plot visualizing crab characteristics with sex and color
}
\description{
Use this function to plot two columns of data using the crabs data set but adds the color and sex as a facet plot.
This plot will show the x and y axis, along with a column selected for data coloring.
With this plot function, two crab characteristics can be visualized along with the differences between sexes and colors.
}
