% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.for.size.R
\name{crabs_model}
\alias{crabs_model}
\title{Function for linear models with crabs data}
\usage{
crabs_model(columnx, columny, data)
}
\arguments{
\item{columnx}{The column for the dependent variable}

\item{columny}{the column for the independent variable}

\item{data}{the data set crabs}
}
\value{
Linear model for comparisons in crabs data
}
\description{
Use this function to produce a linear model for numeric crabs data.
This function will produce a statistical analysis of the data entered into the function and predict correlations.
}
