% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_chr.R
\name{lat_chr}
\alias{lat_chr}
\title{Function to mutate the location from numeric to character}
\usage{
lat_chr(crab_data, column)
}
\arguments{
\item{column}{Column Latitude}

\item{data}{The data set crabs}
}
\value{
New column with Latitude as a character
}
\description{
Use this function to create a new column for the Latitude character and not numeric data.
This function will not delete the location from the data set, but it will select all the columns except the numeric latitude.
}
