% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_fun.R
\name{lm_fun}
\alias{lm_fun}
\title{Function for linear models with crabs data}
\usage{
lm_fun(df, dependent, independent_cols)
}
\arguments{
\item{df}{the data set crabs}

\item{dependent}{The column for the dependent variable}

\item{independent_cols}{the column for the independent variable}
}
\value{
Linear model for comparisons in crabs data
}
\description{
Use this function to produce a linear model for numeric crabs data.
This function will produce a statistical analysis of the data entered into the function and predict correlations.
}
