% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.column.R
\name{select_col}
\alias{select_col}
\title{Function selects and mutates a column.}
\usage{
select_col(data, columns)
}
\arguments{
\item{data}{database being used (data frame)}

\item{columns}{selected columns to mutate}
}
\value{
clean The cleaned and mutated columns with a new mutated column (dataframe)
}
\description{
Use the data to mutate a column as divided by 1000.
Returns a new column with a column divided by 1000 as new_col.
Typically used for mutating a weight column unit.
}
