% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crab_plot.R
\name{crab_plot}
\alias{crab_plot}
\title{Comparison plot for two numeric columns}
\usage{
crab_plot(crab_data, columnx, columny, columnz)
}
\arguments{
\item{columnx}{Column for X value}

\item{columny}{Column for Y value}

\item{columnz}{Column for z value}

\item{data}{The data set crabs}
}
\value{
X-Y-Z plot
}
\description{

}
